+++ src/components/MobileNav/NavList.tsx
@@ -1,8 +1,10 @@
 import "./scss/index.scss";
 
 import * as React from "react";
+import { FormattedMessage } from "react-intl";
 import { Link } from "react-router-dom";
 import ReactSVG from "react-svg";
+import { commonMessages } from "@temp/intl";
 
 import { baseUrl } from "../../app/routes";
 import backImg from "../../images/arrow-back.svg";