+++ src/core/config.ts
@@ -13,8 +13,24 @@ export const PROVIDERS = {
     label: "Dummy",
   },
   STRIPE: {
-    href: "https://js.stripe.com/v3/",
     label: "Stripe",
+  },
+  ADYEN: {
+    label: "Adyen",
+    script: {
+      src:
+        "https://checkoutshopper-test.adyen.com/checkoutshopper/sdk/3.10.1/adyen.js",
+      integrity:
+        "sha384-wG2z9zSQo61EIvyXmiFCo+zB3y0ZB4hsrXVcANmpP8HLthjoQJQPBh7tZKJSV8jA",
+      crossOrigin: "anonymous",
+    },
+    style: {
+      src:
+        "https://checkoutshopper-test.adyen.com/checkoutshopper/sdk/3.10.1/adyen.css",
+      integrity:
+        "sha384-8ofgICZZ/k5cC5N7xegqFZOA73H9RQ7H13439JfAZW8Gj3qjuKL2isaTD3GMIhDE",
+      crossOrigin: "anonymous",
+    },
   },
 };
 export const STATIC_PAGES = [
@@ -58,6 +74,7 @@ export enum CheckoutStep {
   Shipping,
   Payment,
   Review,
+  PaymentConfirm,
 }
 export const CHECKOUT_STEPS = [
   {
@@ -91,5 +108,13 @@ export const CHECKOUT_STEPS = [
     nextActionName: "Place order",
     onlyIfShippingRequired: false,
     step: CheckoutStep.Review,
+  },
+  {
+    index: 4,
+    link: "/checkout/payment-confirm",
+    name: "Payment confirm",
+    onlyIfShippingRequired: false,
+    step: CheckoutStep.PaymentConfirm,
+    withoutOwnView: true,
   },
 ];