+++ src/userAccount/views/OrderDetails/Page.tsx
@@ -23,19 +23,8 @@ const extractOrderLines = (lines: OrderDetail_lines[]): any[] => {
   return lines
     .map(line => ({
       quantity: line.quantity,
-      totalPrice: {
-        ...line.unitPrice,
-        currency: line.unitPrice.currency,
-        gross: {
-          amount: line.quantity * line.unitPrice.gross.amount,
-          ...line.unitPrice.gross,
-        },
-        net: {
-          amount: line.quantity * line.unitPrice.net.amount,
-          ...line.unitPrice.net,
-        },
-      },
-      variant: line.variant,
+      totalPrice: line.totalPrice,
+      ...line.variant,
       name: line.productName,
     }))
     .sort((a, b) =>