+++ src/views/Category/View.tsx
@@ -5,7 +5,7 @@ import { StringParam, useQueryParam } from "use-query-params";
 
 import { prodListHeaderCommonMsg } from "@temp/intl";
 import { IFilters } from "@types";
-
+import { Loader } from "@components/atoms";
 import { MetaWrapper, NotFound, OfflinePlaceholder } from "../../components";
 import NetworkStatus from "../../components/NetworkStatus";
 import { PRODUCTS_PER_PAGE } from "../../core/config";
@@ -13,7 +13,6 @@ import {
   convertSortByFromString,
   convertToAttributeScalar,
   getGraphqlIdFromDBId,
-  maybe,
 } from "../../core/utils";
 import Page from "./Page";
 import {
@@ -143,6 +142,10 @@ export const View: React.FC<ViewProps> = ({ match }) => {
           loaderFull
         >
           {categoryData => {
+            if (categoryData.loading) {
+              return <Loader />;
+            }
+
             if (categoryData.data && categoryData.data.category === null) {
               return <NotFound />;
             }
@@ -151,16 +154,17 @@ export const View: React.FC<ViewProps> = ({ match }) => {
               return <OfflinePlaceholder />;
             }
 
-            const canDisplayFilters = maybe(
-              () =>
-                !!categoryData.data.attributes.edges &&
-                !!categoryData.data.category.name,
-              false
-            );
+            const canDisplayFilters =
+              !!categoryData.data?.attributes?.edges &&
+              !!categoryData.data?.category?.name;
 
             return (
               <TypedCategoryProductsQuery variables={variables}>
                 {categoryProducts => {
+                  if (!canDisplayFilters && categoryProducts.loading) {
+                    return <Loader />;
+                  }
+
                   if (canDisplayFilters) {
                     const handleLoadMore = () =>
                       categoryProducts.loadMore(
@@ -219,6 +223,8 @@ export const View: React.FC<ViewProps> = ({ match }) => {
                       </MetaWrapper>
                     );
                   }
+
+                  return null;
                 }}
               </TypedCategoryProductsQuery>
             );