+++ CHANGELOG.md
@@ -24,6 +24,13 @@ All notable, unreleased changes to this project will be documented in this file.
 - Add test tags to address book - #847 by @dominik-zeglen
 - Refresh user data if mutation fails - #854 by @dominik-zeglen
 - Fix product list data overfetching - #855 by @dominik-zeglen
+- Add Adyen payment gateway - #845 by @orzechdev
+- Fix crash on null price range - #875 by @orzechdev
+- Update collection products query - #879 by @orzechdev
+- Fix checkout refreshing - #865 by @orzechdev
+- Add purchase availability to product details page - #878 by @orzechdev
+- Require payment recreate when payment price is wrong - #892 by @orzechdev
+- Handle JWT token refreshing and verifying - #883 by @orzechdev
 
 ## 2.10.4